requirements.txt
# Core FastAPI and web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP client
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Graph processing
networkx==3.2.1

# Data models
pydantic==2.4.2

# Scheduling
APScheduler==3.10.4

# Document processing
PyPDF2==3.0.1
PyMuPDF==1.23.8
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2

# Report generation
reportlab==4.0.7
weasyprint==60.2

# Templating
jinja2==3.1.2

# Authentication
pyjwt==2.8.0
bcrypt==4.1.1
email-validator==2.0.0.post2
python-jose==3.3.0

# Google services
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# AI/LLM services
openai==1.3.8
google-generativeai==0.3.2
anthropic==0.7.8

# Image processing
Pillow==10.1.0

# MongoDB dependencies
motor==3.3.2
pymongo==4.6.0
certifi==2023.11.17

# Additional dependencies
dnspython==2.4.2
pyotp==2.9.0
qrcode==7.4.2
cryptography>=41.0.0
nanoid==2.0.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Networking and web utilities
starlette==0.27.0
